Add onboard schools to support settings page
Also moves the location of the tests to a more appropriate folder
