+++ app/views/claims/support/settings/index.html.erb
@@ -11,6 +11,8 @@
         # Uncomment the following line and add a translation to add a link to the feature flags page
         # govuk_link_to(t(".feature_flags"), "/flipper", no_visited_state: true, new_tab: true),
         govuk_link_to(t(".claim_windows"), claims_support_claim_windows_path, no_visited_state: true),
+        govuk_link_to(t(".onboard_users"), new_upload_users_claims_support_schools_path, no_visited_state: true),
+        govuk_link_to(t(".onboard_schools"), new_onboard_schools_claims_support_schools_path, no_visited_state: true),
         govuk_link_to(t(".emails"), claims_support_mailers_path, no_visited_state: true),
         (govuk_link_to(t(".reset_database"), reset_claims_support_database_path, no_visited_state: true) unless HostingEnvironment.env.production?),
         (govuk_link_to(t(".revert_claims_to_submitted"), revert_to_submitted_claims_support_database_databases_claims_path, no_visited_state: true) unless HostingEnvironment.env.production?),
