+++ config/locales/en/claims/support/settings.yml
@@ -10,3 +10,4 @@ en:
           reset_database: Reset database
           revert_claims_to_submitted: Revert all claims to 'Submitted'
           onboard_users: Onboard users
+          onboard_schools: Onboard schools
