+++ spec/system/claims/support/settings/onboard_multiple_schools/support_user_can_not_onboard_schools_when_there_are_no_claim_windows_spec.rb
@@ -6,7 +6,8 @@ RSpec.describe "Support user can not onboard schools when there are no claim win
   scenario do
     given_schools_exist
     and_i_am_signed_in
-    when_i_navigate_to_onboard_multiple_schools
+    when_i_navigate_to_the_settings_index_page
+    and_i_navigate_to_onboard_multiple_schools
     then_i_see_the_no_claim_window_page
   end
 
@@ -22,14 +23,14 @@ RSpec.describe "Support user can not onboard schools when there are no claim win
     sign_in_claims_support_user
   end
 
-  def when_i_navigate_to_the_organisations_index_page
+  def when_i_navigate_to_the_settings_index_page
     within(primary_navigation) do
-      click_on "Organisations"
+      click_on "Settings"
     end
   end
 
-  def when_i_navigate_to_onboard_multiple_schools
-    visit new_onboard_schools_claims_support_schools_path
+  def and_i_navigate_to_onboard_multiple_schools
+    click_on "Onboard schools"
   end
 
   def then_i_see_the_no_claim_window_page
