+++ spec/system/claims/support/settings/onboard_multiple_schools/support_user_onboards_multiple_schools_when_only_a_future_claim_window_exists_spec.rb
@@ -11,7 +11,8 @@ RSpec.describe "Support user onboards multiple schools when only a upcoming clai
     given_a_upcoming_claim_window_exists
     and_schools_exist
     and_i_am_signed_in
-    when_i_navigate_to_onboard_multiple_schools
+    when_i_navigate_to_the_settings_index_page
+    and_i_navigate_to_onboard_multiple_schools
     then_i_see_the_claim_window_page
 
     when_i_select_the_upcoming_claim_window
@@ -49,14 +50,14 @@ RSpec.describe "Support user onboards multiple schools when only a upcoming clai
     sign_in_claims_support_user
   end
 
-  def when_i_navigate_to_the_organisations_index_page
+  def when_i_navigate_to_the_settings_index_page
     within(primary_navigation) do
-      click_on "Organisations"
+      click_on "Settings"
     end
   end
 
-  def when_i_navigate_to_onboard_multiple_schools
-    visit new_onboard_schools_claims_support_schools_path
+  def and_i_navigate_to_onboard_multiple_schools
+    click_on "Onboard schools"
   end
 
   def then_i_see_the_upload_page
