+++ spec/system/claims/support/settings/onboard_multiple_schools/support_user_uploads_a_file_with_the_wrong_headers_spec.rb
@@ -5,7 +5,8 @@ RSpec.describe "Support user does upload a file with wrong headers", service: :c
     given_claim_windows_exist
     and_schools_exist
     and_i_am_signed_in
-    when_i_navigate_to_onboard_multiple_schools
+    when_i_navigate_to_the_settings_index_page
+    and_i_navigate_to_onboard_multiple_schools
     then_i_see_the_claim_window_page
 
     when_i_select_the_upcoming_claim_window
@@ -34,14 +35,14 @@ RSpec.describe "Support user does upload a file with wrong headers", service: :c
     sign_in_claims_support_user
   end
 
-  def when_i_navigate_to_the_organisations_index_page
+  def when_i_navigate_to_the_settings_index_page
     within(primary_navigation) do
-      click_on "Organisations"
+      click_on "Settings"
     end
   end
 
-  def when_i_navigate_to_onboard_multiple_schools
-    visit new_onboard_schools_claims_support_schools_path
+  def and_i_navigate_to_onboard_multiple_schools
+    click_on "Onboard schools"
   end
 
   def then_i_see_the_upload_page
